requirements.txt
pygame
opencv-python
vgamepad
pillow
mss
numpy
keras
scikit-learn
pandas
tensorflow